change src

+++ pisces/models/account.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey
 from model import Base
 
 from models.user import User
@@ -11,7 +11,7 @@ class Account(Base):
 
     name = Column(String(64), primary_key = True)
     pwd = Column(String(64))
-    userid = Column(Integer)
+    userid = Column(BigInteger)
 
     def __init__(self, name, pwd, userid):
         self.name = name
