+++ pisces/models/character.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey
 from model import Base
 
 from models.job import Job
@@ -10,8 +10,8 @@ from models.level import Level
 class Character(Base):
     __tablename__ = "character"
 
-    id = Column(Integer, primary_key = True)
+    id = Column(BigInteger, primary_key = True)
     name = Column(String(64))
     gender = Column(String(64))
-    jobid = Column(Integer)
-    levelid = Column(Integer)
+    jobid = Column(BigInteger)
+    levelid = Column(BigInteger)
