+++ pisces/models/level.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey
 from model import Base
 
 class Level(Base):
     __tablename__ = "level"
 
-    id = Column(Integer, primary_key = True)
+    id = Column(BigInteger, primary_key = True)
     level = Column(Integer)
     exp = Column(Integer)
