+++ pisces/models/user.py
@@ -1,7 +1,7 @@
 #!/user/bin/env python 
 # -*- coding:utf-8 -*-
 
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey
 from model import Base
 
 from models.character import Character
@@ -9,6 +9,6 @@ from models.character import Character
 class User(Base):
     __tablename__ = "user"
 
-    id = Column(Integer, primary_key = True)
+    id = Column(BigInteger, primary_key = True)
     name = Column(String(64))
-    characterid = Column(Integer)
+    characterid = Column(BigInteger)
